chore: rename content-script-2 folder
