--- src/content-script-2/components/index.ts
@@ -1,3 +1,0 @@
-export { Overlay } from './overlay';
-export { Spinner } from './spinner';
-export { TextInput } from './text-input';
