--- src/content-script-2/context/index.ts
@@ -1,3 +1,0 @@
-export { SessionContext, withSessionContext } from './session-context';
-export { ActiveScopeContext, withActiveScopeContext } from './active-scope-context';
-export { AgentContext, withAgentContext } from './agent-context';
