--- src/content-script-2/index.ts
@@ -1,1 +1,0 @@
-export { App } from './app';
