--- src/content-script-2/pages/manual-test-error/manual-test-error.module.scss
@@ -1,27 +1,0 @@
-@import "../../../common/style-guide";
-
-.manual-test-error {
-  &__icon {
-    color: map-get($primary-blue, default);
-    cursor: pointer;
-  }
-
-  &__icon-tooltip {
-    width: 180px;
-    top: 50%;
-    right: 0;
-    transform: translateY(-50%);
-    background: linear-gradient(90deg, rgba(255,255,255,0) 0%, rgba(255,255,255,1) 35%, rgba(255,255,255,1) 100%);
-    color: map-get($monochrome, black);
-    font-size: 12px;
-    text-align: right;
-  }
-
-  &__error-message {
-    display: inline-block;
-    max-width: 240px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-}
