+++ src/content-script/app.module.scss
