+++ src/content-script/components/index.ts
@@ -1,0 +1,3 @@
+export { Overlay } from './overlay';
+export { Spinner } from './spinner';
+export { TextInput } from './text-input';
