+++ src/content-script/components/overlay/index.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import { BEM } from '@redneckz/react-bem-helper';
+
+import styles from './overlay.module.scss';
+
+interface Props {
+  className?: string;
+  children?: React.ReactNode;
+}
+
+export const Overlay = BEM(styles)(({ children, className }: Props) => <div className={className}>{children}</div>);
