+++ src/content-script/components/spinner/index.tsx
