+++ src/content-script/components/spinner/spinner.module.scss
