+++ src/content-script/components/text-input/text-input.module.scss
