+++ src/content-script/context/active-scope-context.tsx
